+++ package.json
@@ -17,10 +17,10 @@
     "bs58": "^5.0.0",
     "buffer": "^6.0.3",
     "ee-typed": "^0.1.1",
-    "emp-wasm-backend": "^0.4.0",
+    "emp-wasm-engine": "^0.5.2",
     "events": "^3.3.0",
     "keccak256": "^1.0.6",
-    "mpc-framework": "^0.2.1",
+    "mpc-framework": "^0.3.0",
     "msgpackr": "^1.11.2",
     "peerjs": "^1.5.4",
     "qrcode.react": "^3.1.0",
@@ -29,7 +29,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.0",
     "rtc-pair-socket": "^0.1.7",
-    "summon-ts": "^0.5.3",
+    "summon-ts": "^0.6.0",
     "typed-emitter": "^2.1.0",
     "zod": "^3.23.5"
   },
